requirements.txt
opencv-python
mediapipe
numpy
pillow
scipy
scikit-image
matplotlib
onnxruntime-gpu
torch>=2.1.0
torchvision
lpips
tqdm
pandas
pyyaml